job post add working

+++ JobPostIJP/src/main/java/com/ukgG3/JobPosting/service/impl/JobPostingServiceImpl.java
@@ -1,13 +1,10 @@
 package com.ukgG3.JobPosting.service.impl;
 
 // JobPostingServiceImpl.jav
-import com.ukgG3.JobPosting.entity.CodingLanguage;
 import com.ukgG3.JobPosting.entity.JobPosting;
-import com.ukgG3.JobPosting.dto.EmployeeDto;
 import com.ukgG3.JobPosting.dto.JobPostingDto;
-import com.ukgG3.JobPosting.entity.Location;
-import com.ukgG3.JobPosting.mapper.EmployeeMapper;
 import com.ukgG3.JobPosting.mapper.JobPostingMapper;
+import com.ukgG3.JobPosting.repository.CodingLanguageRepository;
 import com.ukgG3.JobPosting.repository.JobPostingRepository;
 import com.ukgG3.JobPosting.repository.LocationRepository;
 import com.ukgG3.JobPosting.service.IJobPostingService;
@@ -15,7 +12,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
@@ -25,41 +21,33 @@ public class JobPostingServiceImpl implements IJobPostingService {
     @Autowired
     private LocationRepository locationRepository;
     @Autowired
-    private LocationRepository codingLanguageRepository;
-
+    private CodingLanguageRepository codingLanguageRepository;
 
     @Override
     public List<JobPostingDto> getAllJobPostings() {
         return jobPostingRepository.findAll().stream()
-                .map(jobPosting -> {
-                    return JobPostingMapper.mapToJobPostingDto(jobPosting, new JobPostingDto());
-                })
+                .map(JobPostingMapper::mapToJobPostingDto)
                 .collect(Collectors.toList());
     }
 
     @Override
     public JobPostingDto getJobPostingById(Long id) {
         return jobPostingRepository.findById(id)
-                .map(jobPosting -> {
-                    return JobPostingMapper.mapToJobPostingDto(jobPosting, new JobPostingDto());
-                })
+                .map(JobPostingMapper::mapToJobPostingDto)
                 .orElse(null);
     }
 
-
     @Override
-    public boolean saveJobPosting(JobPostingDto jobPostingDTO) {
-        JobPosting jobPosting = JobPostingMapper.mapToJobPosting(jobPostingDTO, new JobPosting());
-        Set<Location> locations = locationRepository.findAllById(jobPostingDTO.getLocations());
-        Set<CodingLanguage> codingLanguages = codingLanguageRepository.findAllById(jobPostingDTO.getCodingLanguages());
-        jobPosting.setLocations(locations);
-        jobPosting.setCodingLanguages(codingLanguages);
+    public boolean saveJobPosting(JobPostingDto dto) {
+        JobPosting jobPosting = JobPostingMapper.mapToJobPosting(dto, locationRepository, codingLanguageRepository);
         jobPostingRepository.save(jobPosting);
         return true;
     }
 
     @Override
     public void deleteJobPosting(Long id) {
-        jobPostingRepository.deleteById(id);
+        if (jobPostingRepository.existsById(id)) {
+            jobPostingRepository.deleteById(id);
+        }
     }
 }
